Update Card.tsx
Repositioned the item hover menu buttons. The "mark as done" button was moved closer to the delete button because it's less frequently used.

The most used button is the edit button, which is now in the middle of the menu (farther away from the delete button). This should make it less likely to accidentally click the delete button when you're editing items.